+++ CHANGELOG.md
@@ -8,9 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-* A new Publication `SearchService` to search through the resources' content, with two default implementations:
-    * `IcuSearchService` uses [ICU components](https://developer.android.com/reference/kotlin/android/icu/text/StringSearch) to take into account language specificities.
-    * `NaiveSearchService` performs exact matches on strings but cannot handle case and diacritic sensitivity.
+* A new Publication `SearchService` to search through the resources' content, with a default implementation `StringSearchService`.
 
 
 ## [2.0.0]